backend/stock-service/pyproject.toml
[tool.poetry]
name = "stock-service"
version = "1.0.0"
description = "Real-time stock market data and websocket service"
authors = ["JSh4w <jontyshaw@btinternet.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "0.104.0"
uvicorn = "0.23.2"
pandas = "2.1.1"
numpy = "1.26.0"
websockets = "15.0.1"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"
httpx = "^0.28.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.0"
uvicorn = {extras = ["standard"], version = "0.23.2"}
watchfiles = "^0.21.0"

[tool.pylint.MASTER]
init-hook = "import sys; sys.path.append('.')"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"